+++ Source/AccelByteUe4Sdk/Public/Core/AccelByteIdValidator.h
@@ -26,5 +26,6 @@ public:
 	static FString GetChannelIdInvalidMessage(FString const& ChannelId);
 	static FString GetContentIdInvalidMessage(FString const& ContentId);
 	static FString GetScreenshotIdInvalidMessage(FString const& ScreenshotId);
+	static FString GetEntitlementIdInvalidMessage(FString const& EntitlementId);
 };
 }